back_end/requirements.txt
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.1
django-crontab==0.7.1
django-environ==0.8.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
libclang==12.0.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.4
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.23
protobuf==3.19.1
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.6.0
